+++ hermes-client/src/test/groovy/pl/allegro/tech/hermes/client/HermesClientMetricsTest.groovy
@@ -14,8 +14,8 @@ import static pl.allegro.tech.hermes.client.HermesClientBuilder.hermesClient
 
 class HermesClientMetricsTest extends Specification {
 
-    private final MetricRegistry metrics = new MetricRegistry()
-    private final MetricsProvider metricsProvider = new DropwizardMetricsProvider(metrics)
+    private MetricRegistry metrics = new MetricRegistry()
+    private MetricsProvider metricsProvider = new DropwizardMetricsProvider(metrics)
 
     def "should measure publish latency"() {
         given: