+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/consumer/batch/DirectBufferUtils.java
@@ -15,11 +15,7 @@ class DirectBufferUtils {
     static {
         DirectBufferCleaner cleaner;
         try {
-            if ("1.8".equals(System.getProperty("java.specification.version"))) {
-                cleaner = new Java8DirectBufferCleaner();
-            } else {
-                cleaner = new Java11DirectBufferCleaner();
-            }
+            cleaner = new Java11DirectBufferCleaner();
         } catch (ReflectiveOperationException e) {
             cleaner = null;
         }
@@ -38,24 +34,6 @@ class DirectBufferUtils {
 
     private interface DirectBufferCleaner {
         void clean(ByteBuffer buffer) throws ReflectiveOperationException;
-    }
-
-    static class Java8DirectBufferCleaner implements DirectBufferCleaner {
-        private final Method directBufferCleaner;
-        private final Method directBufferCleanerClean;
-
-        Java8DirectBufferCleaner() throws ReflectiveOperationException {
-            directBufferCleaner = Class.forName("sun.nio.ch.DirectBuffer").getMethod("cleaner");
-            directBufferCleaner.setAccessible(true);
-            directBufferCleanerClean = Class.forName("sun.misc.Cleaner").getMethod("clean");
-            directBufferCleanerClean.setAccessible(true);
-        }
-
-        @Override
-        public void clean(ByteBuffer buffer) throws ReflectiveOperationException {
-            Object cleaner = directBufferCleaner.invoke(buffer);
-            directBufferCleanerClean.invoke(cleaner);
-        }
     }
 
     static class Java11DirectBufferCleaner implements DirectBufferCleaner {