+++ hermes-frontend/src/test/groovy/pl/allegro/tech/hermes/frontend/publishing/preview/MessagePreviewFactoryTest.groovy
@@ -16,7 +16,7 @@ class MessagePreviewFactoryTest extends Specification {
         given:
             factory = new MessagePreviewFactory(maxContentSize)
         when:
-            MessagePreview preview = factory.create(new JsonMessage('message-id', new byte[messageSize], 0L, "partition-key"))
+            MessagePreview preview = factory.create(new JsonMessage('message-id', new byte[messageSize], 0L, "partition-key"), false)
         then:
             preview.truncated == shouldTruncate
         where:
@@ -34,7 +34,7 @@ class MessagePreviewFactoryTest extends Specification {
             def message = new AvroMessage('message-id', avroUser.asBytes(), 0L, avroUser.compiledSchema, null)
             factory = new MessagePreviewFactory(avroUser.asJson().length() - 1)
         when:
-            MessagePreview preview = factory.create(message)
+            MessagePreview preview = factory.create(message, false)
         then:
             preview.truncated
     }