+++ hermes-frontend/src/test/groovy/pl/allegro/tech/hermes/frontend/publishing/preview/MessagePreviewLogTest.groovy
@@ -1,10 +1,10 @@
 package pl.allegro.tech.hermes.frontend.publishing.preview
 
-import pl.allegro.tech.hermes.domain.topic.preview.MessagePreview
+import pl.allegro.tech.hermes.common.message.wrapper.SchemaAwareSerDe
 import pl.allegro.tech.hermes.frontend.publishing.avro.AvroMessage
 import pl.allegro.tech.hermes.frontend.publishing.message.JsonMessage
-import pl.allegro.tech.hermes.frontend.publishing.message.Message
 import pl.allegro.tech.hermes.test.helper.avro.AvroUser
+import pl.allegro.tech.hermes.test.helper.builder.TopicBuilder
 import spock.lang.Specification
 
 import java.util.concurrent.CountDownLatch
@@ -16,16 +16,12 @@ import static pl.allegro.tech.hermes.api.TopicName.fromQualifiedName
 
 class MessagePreviewLogTest extends Specification {
 
-    private MessagePreviewFactory factory = Stub(MessagePreviewFactory) {
-        create(_ as Message) >> { args -> new MessagePreview(args[0].data) }
-    }
-
-    private MessagePreviewLog log = new MessagePreviewLog(factory, 2)
+    private MessagePreviewLog log = new MessagePreviewLog(new MessagePreviewFactory(10 * 1024), 2)
 
     def "should persist JSON messages for topics"() {
         given:
-        log.add(fromQualifiedName('group.topic-1'), new JsonMessage('id', [1] as byte[], 0L, "partition-key"))
-        log.add(fromQualifiedName('group.topic-2'), new JsonMessage('id', [2] as byte[], 0L, null))
+        log.add(TopicBuilder.topic('group.topic-1').build(), new JsonMessage('id', [1] as byte[], 0L, "partition-key"))
+        log.add(TopicBuilder.topic('group.topic-2').build(), new JsonMessage('id', [2] as byte[], 0L, null))
 
         when:
         def messages = log.snapshotAndClean()
@@ -39,20 +35,42 @@ class MessagePreviewLogTest extends Specification {
         def avroUser = new AvroUser()
         def message = new AvroMessage('message-id', avroUser.asBytes(), 0L, avroUser.compiledSchema, null)
 
-        log.add(fromQualifiedName('group.topic-1'), message)
+        log.add(TopicBuilder.topic('group.topic-1').build(), message)
 
         when:
         def messages = log.snapshotAndClean()
 
         then:
         messages.topics() as Set == [fromQualifiedName('group.topic-1')] as Set
+
+        def preview = messages.previewOf(fromQualifiedName('group.topic-1')).get(0).content
+        def previewString = new String(preview)
+        previewString == """{"__metadata":null,"name":"defaultName","age":0,"favoriteColor":"defaultColor"}"""
+    }
+
+    def "should persist Avro messages for schema aware topics"() {
+        given:
+        def avroUser = new AvroUser()
+        def message = new AvroMessage('message-id', SchemaAwareSerDe.serialize(avroUser.compiledSchema.id, avroUser.asBytes()), 0L, avroUser.compiledSchema, null)
+
+        log.add(TopicBuilder.topic('group.topic-1').withSchemaIdAwareSerialization().build(), message)
+
+        when:
+        def messages = log.snapshotAndClean()
+
+        then:
+        messages.topics() as Set == [fromQualifiedName('group.topic-1')] as Set
+
+        def preview = messages.previewOf(fromQualifiedName('group.topic-1'))
+        def previewString = new String(preview.get(0).content)
+        previewString == """{"__metadata":null,"name":"defaultName","age":0,"favoriteColor":"defaultColor"}"""
     }
 
     def "should persist no more than two messages for topic"() {
         given:
-        log.add(fromQualifiedName('group.topic-1'), new JsonMessage('id', [1] as byte[], 0L, null))
-        log.add(fromQualifiedName('group.topic-1'), new JsonMessage('id', [2] as byte[], 0L, null))
-        log.add(fromQualifiedName('group.topic-1'), new JsonMessage('id', [3] as byte[], 0L, null))
+        log.add(TopicBuilder.topic('group.topic-1').build(), new JsonMessage('id', [1] as byte[], 0L, null))
+        log.add(TopicBuilder.topic('group.topic-1').build(), new JsonMessage('id', [2] as byte[], 0L, null))
+        log.add(TopicBuilder.topic('group.topic-1').build(), new JsonMessage('id', [3] as byte[], 0L, null))
 
         when:
         def messages = log.snapshotAndClean()
@@ -70,7 +88,7 @@ class MessagePreviewLogTest extends Specification {
         threads.times {
             int executor = it
             executorService.submit({
-                1000.times { log.add(fromQualifiedName("group.topic"), new JsonMessage('id', [executor, it] as byte[], 0L, null)) }
+                1000.times { log.add(TopicBuilder.topic("group.topic").build(), new JsonMessage('id', [executor, it] as byte[], 0L, null)) }
                 latch.countDown()
             })
         }