+++ integration/src/integration/java/pl/allegro/tech/hermes/integration/env/FrontendStarter.java
@@ -21,6 +21,8 @@ import static com.jayway.awaitility.Duration.TEN_SECONDS;
 import static javax.ws.rs.core.Response.Status.OK;
 import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_FORCE_TOPIC_MAX_MESSAGE_SIZE;
 import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_GRACEFUL_SHUTDOWN_ENABLED;
+import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_MESSAGE_PREVIEW_ENABLED;
+import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_MESSAGE_PREVIEW_LOG_PERSIST_PERIOD;
 import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_PORT;
 import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_SSL_ENABLED;
 import static pl.allegro.tech.hermes.common.config.Configs.FRONTEND_THROUGHPUT_FIXED_MAX;
@@ -68,6 +70,8 @@ public class FrontendStarter implements Starter<ConfigurableApplicationContext> 
         args.add(getArgument(FRONTEND_THROUGHPUT_TYPE, "fixed"));
         args.add(getArgument(FRONTEND_THROUGHPUT_FIXED_MAX, 50 * 1024L));
         args.add(getArgument(FRONTEND_GRACEFUL_SHUTDOWN_ENABLED, false));
+        args.add(getArgument(FRONTEND_MESSAGE_PREVIEW_ENABLED, true));
+        args.add(getArgument(FRONTEND_MESSAGE_PREVIEW_LOG_PERSIST_PERIOD, 1));
         return args;
     }